fix(server): bigint migration from the start, failed runs rank in worst sort, lastReadyRun in enriched list

+++ packages/server/drizzle/pg/0018_modern_rocket_racer.sql
@@ -1,3 +1,3 @@
-ALTER TABLE "runs" ADD COLUMN "duration_ms" integer;
+ALTER TABLE "runs" ADD COLUMN "duration_ms" bigint;
 --> statement-breakpoint
 UPDATE runs SET duration_ms = NULLIF((stats_json::jsonb ->> 'durationMs'), '')::bigint WHERE stats_json IS NOT NULL;
